🚑 Fix defaultPlayer and ArcheAnimation to avoid animations and controls bugs. ✅ Update and fix RacingVehicle !✅ Update charactersData to consider differents controllers configurations. ✅ Rename Character to 'DefaultCharacter', it's considered as a Arche controller configuration.
