+++ src/app/playground/core/character/character.type.ts
@@ -2,5 +2,5 @@ export type CharacterProps = {
     name: string;
     position?: [number, number, number];
     defaultPlayer?: boolean;
-    path: string;
+    path?: string;
   };
