+++ src/app/playground/core/extension/arche/Arche.tsx
@@ -297,6 +297,9 @@ const ARCHE: ForwardRefRenderFunction<customRigidBody, ArcheProps> = (
     action4: 0,
   };
 
+  // If defaultPlayer is true, the character will be dynamic.
+  const characterType = defaultPlayer ? "dynamic" : "fixed";
+
   const characterRigidBody = useWorldRigidBody(characterRef);
 
   /** Move and Camera mode */
@@ -1565,6 +1568,7 @@ const ARCHE: ForwardRefRenderFunction<customRigidBody, ArcheProps> = (
   return (
     <RigidBody
       ref={characterRef}
+      type={characterType}
       position={props.position || [0, 5, 0]}
       friction={props.friction || -0.5}
       mass={props.mass || 100}
